changes on subtract local cart items

+++ Fastkart-Front-SSR/src/app/shared/components/widgets/product-box/product-box-horizontal/product-box-horizontal.component.ts
@@ -7,7 +7,7 @@ import { Product } from '../../../../../shared/interface/product.interface';
 import { CartAddOrUpdate, Cart } from '../../../../../shared/interface/cart.interface';
 import { AddOrRemoveWishlist, DeleteWishlist } from '../../../../../shared/action/wishlist.action';
 import { AddToCompare } from '../../../../../shared/action/compare.action';
-import { AddToCart, AddToCartLocalStorage, DeleteCart, GetCartItems } from '../../../../../shared/action/cart.action';
+import { AddToCart, AddToCartLocalStorage, DeleteCart, GetCartItems, SubractFromCartLocalStorage } from '../../../../../shared/action/cart.action';
 import { CartState } from '../../../../../shared/state/cart.state';
 import { VariationModalComponent } from '../../modal/variation-modal/variation-modal.component';
 import { environment } from 'src/environments/environment';
@@ -122,11 +122,10 @@ export class ProductBoxHorizontalComponent {
       responseObject.qty = 1;
     }
 
-    if(userId){
+    if (userId) {
       this.store.dispatch(new AddToCart(responseObject));
-    }else{
-      product.qty=responseObject.qty;
-      console.log(product)
+    } else {
+      product.qty = responseObject.qty;
       this.store.dispatch(new AddToCartLocalStorage(product));
     }
   };
@@ -144,15 +143,25 @@ export class ProductBoxHorizontalComponent {
     const itemFound = this.cartItems.find((item: any) => item.product_id === product.id);
     const formattedQty = convertStringToNumber(itemFound.qty);
     if ((formattedQty - 1) === 0) {
+      product.qty = 0;
       const object = {
         id: itemFound.id
       };
 
-      this.store.dispatch(new DeleteCart(object));
+      if (userId) {
+        this.store.dispatch(new DeleteCart(object));
+      } else {
+        this.store.dispatch(new SubractFromCartLocalStorage(product));
+      }
     } else {
       requestObject.qty = formattedQty - 1;
 
-      this.store.dispatch(new AddToCart(requestObject));
+      if (userId) {
+        this.store.dispatch(new AddToCart(requestObject));
+      } else {
+        product.qty = requestObject.qty;
+        this.store.dispatch(new SubractFromCartLocalStorage(product));
+      }
     }
   };
